Moves againstmice class into lib, adds into new branch
